+++ src/game/tests.rs
@@ -353,6 +353,7 @@ mod vintage_cube_cut_down;
 mod vintage_cube_dack;
 mod vintage_cube_damn;
 mod vintage_cube_dark_confidant;
+mod vintage_cube_dauthi_voidwalker;
 mod vintage_cube_delayed_blast_fireball;
 mod vintage_cube_descendant_of_storms;
 mod vintage_cube_detectives_phoenix;